+++ python/demo.py
@@ -15,7 +15,10 @@ from rapidocr import RapidOCR, VisRes
 # yaml_path = "tests/test_files/config.yaml"
 # engine = RapidOCR(config_path=yaml_path)
 
-engine = RapidOCR()
+engine = engine = RapidOCR(
+    params={"Global.with_onnx": True, "EngineConfig.onnxruntime.use_cuda": True}
+)
+# engine = RapidOCR(params={"Global.with_openvino": True})
 vis = VisRes()
 
 image_path = "tests/test_files/ch_en_num.jpg"
