+++ python/rapidocr/config.yaml
@@ -1,6 +1,7 @@
 Global:
     lang: "ch"
     text_score: 0.5
+
     use_det: true
     use_cls: true
     use_rec: true
@@ -60,6 +61,7 @@ Cls:
 
 Rec:
     model_path: models/ch_PP-OCRv4_rec_infer.onnx
+    rec_keys_path: models/ppocr_keys_v1.txt
 
     rec_img_shape: [3, 48, 320]
     rec_batch_num: 6
